project/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.3.3
cattrs==23.2.3
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.57.0
django-cors-headers==4.3.0
django-filter==23.3
django-safedelete==1.4.0
django-shortuuidfield==0.1.3
django-soft-delete==1.0.13
django-storages==1.14.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
dropbox==12.0.2
factory-boy==3.3.0
Faker==20.1.0
filelock==3.13.1
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
httplib2==0.22.0
identify==2.5.34
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jedi==0.19.1
kombu==5.3.7
lsprotocol==2024.0.0a1
msgpack==1.0.8
multimethod==1.11.1
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.2
parso==0.8.3
paymob-solutions==1.0.0
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
pre-commit==3.6.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
pyright==1.1.350
pytest==8.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruff==0.1.15
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.0.6
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
wrapt==1.16.0
